// frontend/src/app/shared/generated/models/search-results-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface SearchResultsRequest {
  fromUtc?: (string | null);
  instrumentId?: (number | string | null);
  sampleIdentifier?: (string | null);
  templateName?: (string | null);
  testId?: (number | string | null);
  toUtc?: (string | null);
}